delete old mods helper classes and move other used classes
