+++ src/main/java/net/sf/jabref/logic/msbib/MSBibConverter.java
@@ -5,8 +5,6 @@ import java.util.List;
 import java.util.Locale;
 
 import net.sf.jabref.logic.layout.format.LatexToUnicodeFormatter;
-import net.sf.jabref.logic.mods.PageNumbers;
-import net.sf.jabref.logic.mods.PersonName;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
