+++ src/main/java/net/sf/jabref/logic/msbib/PersonName.java
@@ -1,4 +1,4 @@
-package net.sf.jabref.logic.mods;
+package net.sf.jabref.logic.msbib;
 
 import java.util.List;
